Put multilang code problem into its own file
